requirements.txt
torch==2.1.2
fastapi==0.99.1
joblib==1.3.2
numpy==1.23.5
Pillow==9.5.0
pydantic==1.10.12
scikit-learn==1.2.0
requests==2.31.0
typing-extensions==4.12.2
json5==0.9.14
uvicorn==0.17.6
gunicorn==22.0.0